@@ -77,12 +54,12 @@ exclude_lines = [
     "raise NotImplementedError",
     "if 0:",
     "if __name__ == .__main__.:",
-    "if TYPE_CHECKING:"
+    "if TYPE_CHECKING:",
 ]
 
 [tool.black]
 line-length = 79
-target-version = ['py38']
+target-version = ["py311"]
 exclude = '''
 /(
     \.eggs
