@@ -1,17 +1,14 @@
-import { simplePost, _getById } from './utils.js'
+import { simplePost, _getById, _elWithAttrs } from './utils'
 
 export function listenForHistoricalReset () {
   const form = _getById('historicalReset')
   form.addEventListener('click', function (e) {
     e.preventDefault()
-    // hide 'done' in case button is pressed again
-    form.querySelector('.done').classList.add('hidden')
-    form.querySelector('.pending').classList.remove('hidden')
-
-    simplePost('reload_historical').then(data => {
-      console.log(`${data}: reload success`)
-      form.querySelector('.pending').classList.add('hidden')
-      form.querySelector('.done').classList.remove('hidden')
+    simplePost('api/historical_reset').then(data => {
+      const resetForm = _getById('historicalReset')
+      resetForm.children[0].disabled = true
+      const notice = _elWithAttrs('h3', { text: 'Historical data reloading...' })
+      resetForm.after(notice)
     }).catch((err) => {
       console.log(`Couldn't reload historical date: ${err}`)
     })
