@@ -1,31 +1,33 @@
-import { _elWithAttrs, getJson, _getById } from '../modules/utils.js'
+import { _getById } from '../modules/utils'
+import { WebsocketClient } from '../modules/websocket_client'
 
-window.addEventListener('DOMContentLoaded', function () {
-  const url = window.location.href + '_update'
-  const dateEl = _getById('date')
-  const seqEl = _getById('seqNum')
-  const eventLink = _getById('eventLink')
-  let currentSeq = parseInt(seqEl.innerText)
-  let currentDate = dateEl.innerText
-  let eventImage = _getById('eventImage')
-  setInterval(function () {
-    getJson(url).then(data => {
-      if (data.seqNum !== currentSeq || data.date !== currentDate) {
-        currentSeq = data.seqNum
-        currentDate = data.date
-        seqEl.textContent = currentSeq
-        dateEl.textContent = currentDate
-        eventLink.href = data.url
-        const newEventImage = _elWithAttrs('img', {
-          id: 'eventImage',
-          class: 'resp',
-          src: data.url
-        })
-        newEventImage.addEventListener('load', (e) => {
-          eventImage.replaceWith(newEventImage)
-          eventImage = _getById('eventImage')
-        })
-      }
+(function () {
+  const initEvent = window.APP_DATA.initEvent || null
+  if (!initEvent) {
+    return
+  }
+  const location = document.documentElement.dataset.locationname
+  const camera = initEvent.camera_name
+  const channel = initEvent.channel_name
+  const baseImgURL = window.APP_DATA.imgURL.split('/').slice(0, -1).join('/')
+  // eslint-disable-next-line no-unused-vars
+  const ws = new WebsocketClient()
+  ws.subscribe('service', 'channel', location, camera, channel)
+
+  window.addEventListener('channel', (message) => {
+    const newEvent = message.detail.data
+    const filename = [newEvent.filename, newEvent.ext].join('.')
+    _getById('date').textContent = newEvent.day_obs
+    _getById('seqNum').textContent = newEvent.seq_num
+    _getById('eventName').textContent = filename
+    const oldImg = _getById('eventImage')
+    const newImg = document.createElement('img')
+    const imgURL = `${baseImgURL}/${filename}`
+    newImg.src = imgURL
+    newImg.id = oldImg.id
+    newImg.addEventListener('load', () => {
+      oldImg.replaceWith(newImg)
+      _getById('eventLink').href = imgURL
     })
-  }, 5000)
-})
+  })
+})()
