@@ -8,7 +8,7 @@ __version__: str
 """The application version string (PEP 440 / SemVer compatible)."""
 
 try:
-    __version__ = version(__name__)
+    __version__ = version("rubintv")
 except PackageNotFoundError:
     # package is not installed
     __version__ = "0.0.0"
