@@ -1,320 +1,262 @@
-from __future__ import annotations
-
+"""Models for rubintv."""
 import re
-from dataclasses import dataclass, field
 from datetime import date, datetime, timedelta
+from typing import Any, Type
 
 from dateutil.tz import gettz
+from pydantic import BaseModel, field_validator
+from pydantic.dataclasses import dataclass
+from typing_extensions import NotRequired, TypedDict
 
-from rubintv.models.models_helpers import string_int_to_date
+__all__ = [
+    "Location",
+    "Channel",
+    "Camera",
+    "Event",
+    "get_current_day_obs",
+]
 
 
-@dataclass
-class Channel:
-    """A container for a category of images from a `Camera`.
-
-    Parameters
-    ----------
-    name : `str`
-        Name of the channel.
-    prefix : `str`
-        Lookup prefix for the GCS Bucket.
-    slug : `str`
-        The part of the URL that identifies this channel in a web request.
-        Simplified version of the name, lowercase with no white-space and only
-        ``-`` or ``_`` seperators.
-    label : `str`
-        Optional extra string for GUI buttons. If not given, the label is set
-        the same as the name.
-    service_dependency : `str`
-        Name of the production service on which the production generator of
-        this channel depends. e.g. ``"auxtel_monitor"`` is dependent on
-        ``"auxtel_isr_runner"``. Used as a url slug (see slug above).
-    """
-
+class Channel(BaseModel):
     name: str
-    prefix: str
-    slug: str = ""
+    title: str
     label: str = ""
-    service_dependency: str = ""
-
-    def __post_init__(self) -> None:
-        if self.label == "":
-            self.label = self.name
+    per_day: bool = False
+    colour: str = ""
 
 
-@dataclass
-class Camera:
-    """A container for a single camera
+class Camera(BaseModel):
+    """Represents a camera entity, capable of handling different channels like
+    images or movies.
 
-    Parameters
+    This class extends the Pydantic BaseModel to leverage data validation. It
+    includes various attributes related to the camera, like its name, online
+    status, and associated channels. It also provides methods for setting
+    default values and categorizing channels.
+
+    Attributes
     ----------
-    name : `str`
-        Name of the camera/image producer.
-    online : `bool`
-        To display the camera or not.
-    slug : `str`
-        The part of the URL that identifies this camera in a web request.
-        Simplified version of the name, lowercase with no white-space and only
-        ``-`` or ``_`` seperators.
-    metadata_slug: `str`
-        Optional slug used for locating shared metadata files in the bucket.
-        If none is given, the slug above is used.
-    logo : `str`
-        Name (including extension) of the image file to display in the GUI
-        button. Images are looked for in ``"/static/images/logos/"``.
-    has_image_viewer : `bool`
-        If ``True`` a column is drawn in the table for this camera with a link
-        generated to point to an external image viewer for each monitor event.
-    channels : `dict` [`str`, `Channel`]
-        `Channel`s belonging to this camera, keyed by name.
-    per_day_channels : `dict` [`str`, `Channel`]
-        `Channel`s belonging to this camera, keyed by name. Per day
-        channels are for categories of event that only occur once per
-        observation date e.g. a movie of the night's images.
-    night_report_prefix : `str`
-        Used to form part of the bucket lookup for night reports. If left unset
-        the camera is considered not to produce night reports.
-    night_report_label : `str`
-        The label used on the link button, page title and subtitle for Night
-        Report-like pages. Defaults to `"Night Report"`
-    metadata_headers : `dict` [`str`, `dict` [`str`, `str`]]] | `dict`[`str`,
-    `str`]
-        The default column headers for the metadata in the table for this
-        `Camera`.
-        -   If it's a simple `dict`, a key is a column name and a value is a
-            description of the column.
-        -   The more complex dict is keyed by group name and then as above.
+    name : str
+        The name of the camera.
+    title : str
+        The title associated with the camera.
+    online : bool
+        Indicates whether the camera is online.
+    metadata_from : str, optional
+        The source of metadata for the camera. Defaults to an empty string and
+        uses `name` if not set.
+    logo : str, optional
+        The logo associated with the camera. Defaults to an empty string.
+    channels : list[Channel], optional
+        A list of channels (either images or movies) associated with the
+        camera. Defaults to an empty list.
+    night_report_label : str, optional
+        Label for the night report. Defaults to "Night Report".
+    metadata_cols : dict[str, str] | None, optional
+        A dictionary defining metadata columns. Defaults to
+        None.
+    image_viewer_link : str, optional
+        A link to the image viewer. Defaults to an empty string.
+    copy_row_template : str, optional
+        Template string for copying a row. Defaults to an empty string.
 
-    js_entry : `str`
-        Used in the template to point to a javascript file that manages table
-        updating and other dynamic interfaces.
+    Methods
+    -------
+    default_as_name(cls: Type, v: str, values: Any) -> str
+        A class method that acts as a field validator for 'metadata_from'. It
+        defaults to using `name` if `metadata_from` is not set.
+    seq_channels() -> list[Channel]
+        Returns a list of sequential channels, i.e., channels that do not have
+        a per-day configuration.
+    pd_channels() -> list[Channel]
+        Returns a list of per-day channels, i.e., channels that have a per-day
+        configuration.
     """
 
     name: str
+    title: str
     online: bool
-    _slug: str = field(init=False, repr=False)
-    metadata_slug: str = ""
+    metadata_from: str = ""
     logo: str = ""
+    channels: list[Channel] = []
+    night_report_label: str = "Night Report"
+    metadata_cols: dict[str, str] | None = None
     image_viewer_link: str = ""
-    channels: dict[str, Channel] = field(default_factory=dict)
-    per_day_channels: dict[str, Channel] = field(default_factory=dict)
-    night_report_prefix: str = ""
-    night_report_label = "Night Report"
-    metadata_headers: dict[str, dict[str, str]] | dict[str, str] = field(
-        init=False, repr=False
-    )
-    js_entry: str = ""
+    copy_row_template: str = ""
 
-    @property
-    def slug(self) -> str:
-        return self._slug
+    # If metadata_from not set, use name as default
+    @field_validator("metadata_from")
+    def default_as_name(cls: Type, v: str, values: Any) -> str:
+        return v or values.get("name")
 
-    @slug.setter
-    def slug(self, slug: str) -> None:
-        self._slug = slug
-        if not self.metadata_slug:
-            self.metadata_slug = slug
+    def seq_channels(self) -> list[Channel]:
+        return [c for c in self.channels if not c.per_day]
+
+    def pd_channels(self) -> list[Channel]:
+        return [c for c in self.channels if c.per_day]
 
 
-@dataclass
-class Location:
-    """Describes a site grouping of cameras
-
-    Parameters
-    ----------
-    name: `str`
-        The name of the location.
-    bucket: `str`
-        The identifying name of the `Bucket` that stores this group's images and
-        other files.
-    services: `list` [`str`]
-        The names of services to display on the admin status page as belonging
-        to this location.
-    slug: `str`
-        The part of the URL that identifies this location in a web request.
-        Simplified version of the name, lowercase with no white-space and only
-        ``-`` or ``_`` seperators.
-    logo: `str`
-        Name (including extension) of the image file to display in the GUI
-        button. Images are looked for in the path ``"/static/images/logos/"``.
-    camera_groups: `dict` [`str`, `list` [`str`]]
-        Used on the location's page. The dict key is used as the group title,
-        and the list is of strings that are keys in a globally accessable dict
-        of `Camera` objects (via
-        ``web.Application["rubintv/models"]["cameras"]``)
-    """
-
+class Location(BaseModel, arbitrary_types_allowed=True):
     name: str
-    bucket: str
-    services: list[str]
-    slug: str = ""
+    title: str
+    bucket_name: str
+    profile_name: str
+    camera_groups: dict[str, list[str]]
+    cameras: list[Camera] = []
     logo: str = ""
-    camera_groups: dict[str, list[str]] = field(default_factory=dict)
-
-    def all_cameras(self) -> list[str]:
-        """Returns list of keys for the camera dict
-
-        Returns
-        -------
-        list[str]
-            _description_
-        """
-        all_cams: list[str] = []
-        for cam_list in self.camera_groups.values():
-            for cam in cam_list:
-                all_cams.append(cam)
-        return all_cams
 
 
 @dataclass
 class Event:
-    """Wrapper for a single unit of GCS `Blob` metadata.
+    key: str
+    hash: str = ""
+    # derived fields:
+    camera_name: str = ""
+    day_obs: str = ""
+    channel_name: str = ""
+    seq_num: int | str = ""
+    filename: str = ""
+    ext: str = ""
 
-    The various fields are extrapolated from the Blob's public url which takes
-    the form:
+    def __post_init__(self) -> None:
+        (
+            self.camera_name,
+            self.day_obs,
+            self.channel_name,
+            self.seq_num,
+            self.filename,
+            self.ext,
+        ) = self.parse_key()
 
-    ``f"{hostname}/{bucket_name}/{channel_prefix}/{channel-dashes-prefix}_dayObs_
-    {date}_seqNum_{seq}.{ext}"``
-    where:
-        -   ``channel-dashes-prefix`` is ``channel_prefix`` with underscores
-            converted to dashes
-        -   ``date`` is as ``"YYYY-MM-DD"``
-        -   ``seq`` is an unpadded integer
-        -   ``ext`` is the file extension i.e. ``.png``
+    def parse_key(self) -> tuple:
+        """Parses a channel event object's key.
 
-    Returns
-    -------
-    url: `str`
-        The public url of the blob.
-    name: `str`
-        The last part of the url after the prefix.
-    prefix: `str`
-        The part of the url that identifies the channel the event belongs to.
-    obs_date: `date`
-        The date of the event.
-    seq: `int`
-        The sequence number of the event.
-        In the case of All Sky, the sequence can be ``"final"`` which
-        is converted to an integer.
-    """
-
-    url: str
-    name: str = field(init=False)
-    prefix: str = field(init=False)
-    obs_date: date = field(init=False)
-    seq: int = field(init=False)
-
-    def parse_filename(self, delimiter: str = "_") -> tuple:
-        """Split the filename into class attributes.
-
-        Parameters
-        ----------
-        delimiter : str, optional
-            The character that delimits the attributes, by default "_"
+        A key is as:
+        ``f"{camera}/{date_str}/{channel}/{seq:06}/{filename}.{ext}"``
 
         Returns
         -------
-        values : `tuple`
-            A tuple of values used by `__post_init__` to fully init the object.
+        url_parts: `tuple`
         """
-        # Every bucket starts /rubintv
-        regex = r"\/rubintv[_\w]+\/"
-        cleaned_up_url = re.split(regex, self.url)[-1]
-        prefix, name = cleaned_up_url.split(
-            "/"
-        )  # We know the name is the last part of the URL
-        nList = name.split(delimiter)
-        the_date = nList[2]
-        year, month, day = map(int, the_date.split("-"))
-        seq_str = nList[4][:-4]  # Strip extension
-        if seq_str == "final":
-            seq = 99999
+        key = self.key
+        name_re = re.compile(
+            r"(\w+)\/([\d-]+)\/(\w+)\/(\d{6}|final)\/([\w-]+)\.(\w+)$"
+        )
+        if match := name_re.match(key):
+            parts = match.groups()
         else:
-            seq = int(seq_str)
-        return (name, prefix, date(year, month, day), seq)
+            raise ValueError(f"Key can't be parsed: {key}")
 
-    def clean_date(self) -> str:
-        return self.obs_date.strftime("%Y-%m-%d")
+        camera, day_obs_str, channel, seq_num, filename, ext = parts
 
-    def __post_init__(self) -> None:
-        self.name, self.prefix, self.obs_date, self.seq = self.parse_filename()
+        try:
+            self.date_str_to_date(day_obs_str)
+        except ValueError:
+            raise ValueError(f"Date can't be parsed: {key}")
+
+        if seq_num != "final":
+            seq_num = int(seq_num)
+
+        return (camera, day_obs_str, channel, seq_num, filename, ext)
+
+    def date_str_to_date(self, date_str: str) -> date:
+        y, m, d = date_str.split("-")
+        return date(int(y), int(m), int(d))
+
+    def day_obs_date(self) -> date:
+        return self.date_str_to_date(self.day_obs)
+
+    def seq_num_force_int(self) -> int:
+        return self.seq_num if isinstance(self.seq_num, int) else -1
 
 
 @dataclass
-class Night_Report_Event:
+class NightReport:
     """Wrapper for a night report blob.
 
         -   Night Reports can be located in a given bucket using the prefix:
-            ``f"/{bucket}/{night_report_prefix}/"``.
+            ``f"/{camera_name}/{date_str}/night_report/"``.
 
-        -   Filenames take the form:
-            ``f"/{obs_date}/{group_name}/{measurement_name}.{file_ext}"``.
+        -   Plots take the form:
+            ``f"/{camera_name}/{date_str}/night_report/{group}/{filename}.{ext}"``.
+
+        -   Metadata takes the form:
+             ``f"/{camera_name}/{date_str}/night_report/{filename}_md.json"``.
+
 
     Parameters
     ----------
-    url: `str`
-        The public url of the blob. This is used as the unique identifier of a
-        plot image onsite.
-    prefix: `str`
-        The prefix used to locate the blob in the bucket. It's used as the
-        delimeter by which to split the url into bucket hostname and the data
-        stored in the filename.
+    key: `str`
+        The name used to find the object in the bucket.
     hash: `str`
         The md5 hash of the blob. This is used to keep plot images up-to-date
         onsite.
-    blobname: `str`
-        The name used to download the blob data from the bucket.
+    camera: `str`
+        The name of the camera the report belongs to.
+    day_obs: `str`
+        A hyphenated string representing the date i.e. ``"2023-12-31"``
     group: `str`
-        The name of the group the data belogs to.
-    name: `str`
-        The name of the plot or text item of data.
-    _obs_date: `str`
-        The observation date of the data.
-    file_ext: `str`
+        The name of the group the data belogs to. In the case the file is
+        metadata, the group is set as ``"metadata"``
+    filename: `str`
+        The fully qualified name of the plot or text item of data.
+    ext: `str`
         The file extension.
     """
 
-    url: str
-    prefix: str
+    key: str
     hash: str
-    blobname: str = ""
-    group: str = field(init=False)
-    name: str = field(init=False)
-    _obs_date: str = field(init=False)
-    file_ext: str = field(init=False)
+    # derived fields
+    camera: str = ""
+    day_obs: str = ""
+    group: str = ""
+    filename: str = ""
+    ext: str = ""
 
-    @property
-    def obs_date(self) -> date:
-        return string_int_to_date(self._obs_date)
-
-    def parse_filename(self) -> tuple:
+    def parse_key(self) -> tuple:
         """Split the filename into parts.
 
         Returns
         -------
         `tuple`
-             A tuple of values used by `__post_init__` to fully init the object.
+             Tuple of values used by `__post_init__` to fully init the object.
         """
-        parts = self.url.split(self.prefix + "/")[-1]
-        # use spread in case of extended names
-        d, group, *names = parts.split("/")
-        obs_date = d
-        raw_name, file_ext = "".join(names).split(".")
-        name = "".join(raw_name).replace("_", " ")
-        return (group, name, obs_date, file_ext)
+        key = self.key
+        metadata_re = re.compile(
+            r"(\w+)\/([\d-]+)\/night_report\/([\w-]+_md)\.(\w+)$"
+        )
+        if match := metadata_re.match(key):
+            parts = match.groups()
+            camera, day_obs_str, filename, ext = parts
+            group = "metadata"
+        else:
+            plot_re = re.compile(
+                r"(\w+)\/([\d-]+)\/night_report\/([\w-]+)\/([\w-]+)\.(\w+)$"
+            )
+            if match := plot_re.match(key):
+                parts = match.groups()
+                camera, day_obs_str, group, filename, ext = parts
+            else:
+                raise ValueError(f"Key can't be parsed: {key}")
+
+        return (camera, day_obs_str, group, filename, ext)
 
     def __post_init__(self) -> None:
         (
+            self.camera,
+            self.day_obs,
             self.group,
-            self.name,
-            self._obs_date,
-            self.file_ext,
-        ) = self.parse_filename()
+            self.filename,
+            self.ext,
+        ) = self.parse_key()
 
-    def dict(self) -> dict:
-        _dict = self.__dict__.copy()
-        return _dict
+    def __hash__(self) -> int:
+        return int(f"0x{self.hash}", 0)
+
+
+class NightReportPayload(TypedDict):
+    text: NotRequired[dict]
+    plots: NotRequired[list[NightReport]]
 
 
 def get_current_day_obs() -> date:
